eric_kudler/barnone/requirements.txt
appnope==0.1.0
backcall==0.1.0
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
fpdf==1.7.2
imutils==0.5.2
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.0
python-barcode==0.9.0
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
pyzbar==0.1.7
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
toolz==0.9.0
traitlets==4.3.2
wcwidth==0.1.7
